Bug 41629: Fix errors with MOZ_SERVICES_SYNC=False
MOZ_SERVICES_SYNC should have been removed in Moz Bug 1227361.
However, it is still available, so we would like to use it.
Since it is a configuration that Mozilla does not test, and using it
results in a build error and in a few runtime errors, too.
This commit fixes them.

We have an upstream bug, too, but its (proposed) fix does not apply to
ESR 102 because of ESMification.
Should it not be accepted, we could replace this commit with the
poposed fix.
Upstream: https://bugzilla.mozilla.org/show_bug.cgi?id=1816969