squash! Bug 40002: Remove about:ion
Bug 41662: Disable about:sync-logs

Even though we disable sync by default with
`identity.fxaccounts.enabled`, this about: page is still avilable.
We could throw an exception on the constructor of the related
component, but it would result only in an error in the console, without
a visible "this address does not look right" error page.
If we fix the issues with MOZ_SERVICES_SYNC, we can restore the
component.


(cherry picked from commit 46dc11a619ebb28f1d63e3d8fe94015718db992a)