fixup! TB 27476: Implement about:torconnect captive portal within Tor Browser
TB 43465: Show the urlbar Connect button when the user might want to
return to about:torconnect.

Instead of hiding the button when canBeginNormalBootstrap is false we
show it as a plain button. Instead we hide it only when we are already
bootstrapped.

We also avoid initialising the button when TorConnect is not enabled.

We also update TorConnect.open:

1. Do not re-open about:torconnect if we are already connected. E.g.
   when the user selects "Connect" in a bridge dialog but we are
   already connected by the time the settings are applied.
2. Do not call TorConnect.startAgain when receiving a "hard" request.
   Only the bridge dialogs make this request, and they would have
   already triggered startAgain by changing the bridge settings.