BB 43544: DoH pane undefined error in Privacy and Security
From: Sarah Jamie Lewis <sarah@openprivacy.ca>
Date: Fri, 28 Feb 2025 09:30:45 -0800
Subject: [PATCH 1/1] DoH Settings: Check for nulll gParentalControlsService

When the parental controls service is disabled in a build, the DoH
settings now display the correct stauts when Increased or Max Protection
is enabled.

Previously, selecting either of these options would cause DoH to be
enabled, but the "Status" and "Provider" fields would not be properly
populated, due to a check on the gParentalControlsService causing an
error.

This check is now identical to the same check in DownloadIntegration.sys.mjs

Apply 1 suggestion(s) to 1 file(s)

Co-authored-by: ma1 <giorgio@maone.net>