fixup! TB 31286: Implementation of bridge, proxy, and firewall settings in about:preferences#connection
TB 43130: Adjust preferences for contrast theme or forced colors.

Give the "Your bridges" and "request bridges" boxes an outline for
forced-colors since the background color will not show.

Use `--border-radius-*` instead of fixed `px` values.

Do not use the accent color for text when in contrast theme or forced
colors.