fixup! BB 40925: Implemented the Security Level component
TB 43782: Update security level UI for new UX flow.

In addition, we drop the distinction between the security levels in the
UI when the user has a custom security level.

I.e. we always show shield as unfilled but with a yellow dot in the
toolbar, and we just call it "Custom" rather than "Standard Custom",
etc.