fixup! Firefox preference overrides.
BB 44026: Disable the modal with upstream's ToU.

We disable data collection, so those terms do not apply to us.
Indeed, currently this feature is already gated on the datareporting
pref, but since upstream is refactoring these prefs, adding an explicit
one seems a good idea.

Also, move the datareporting prefs closer to telemetry.