fixup! [android] Modify build system
TB 43826: Drop an unneeded version customization.

When the patch was originally created
(android-components@abf37979a74f34730dd8e8323d8a9aad184884c3), it
replaced a Date additions that prevented reproducible builds.

However, Mozilla themselves removed that Date in Bug 1824818, so we do
not need this part of the patch anymore.