fixup! [android] Modify build system
TB 43799: Check we passed the objdir to tba-sign-devbuilds.sh.

When we initially created tba-sign-devbuilds.sh, it was easy to find
the APKs. However, after Bug 1951190, they are inside the obj-*
directory, which can be customized in mozconfig.

So, the easiest way to deal with this, was to ask the caller to provide
that obj directory. Not passing it will very likely make the script
fail, so with this commit we added a check to output a more meaningful
error message.