fixup! BB 31740: Remove some unnecessary RemoteSettings instances
BB 43795: Restore the URL classifier XPCOM components.

We are not really interested in removing them.
We're interested in them not calling Mozilla.
However, we have a central patch for RemoteSettings, so we do not need
to patch the single user.
Or, if we did, at least we should patch
UrlClassifierExceptionListService where it creates its RemoteSettings
instance.