fixup! BB 40925: Implemented the Security Level component
BB 44178: Refactor DDG's change of behavior with sec level.

With the highest security level, we use the HTML version of DuckDuckGo.
We used to change the URL on the fly when doing the search, but this
had some problems (different UI, and problems with DDG lite).

With this change, we consider the security level when loading the
search engine configuration, and change the search URLs at that point.